feat(client): split out ISecretAgentClass
