+++ client/interfaces/IAwaitedOptions.ts
@@ -1,5 +1,5 @@
-import CoreTab from "../lib/CoreTab";
-import { ISecretAgent } from "./ISecretAgent";
+import CoreTab from '../lib/CoreTab';
+import ISecretAgent from './ISecretAgent';
 
 export default interface IAwaitedOptions {
   secretAgent: ISecretAgent;
