+++ client/interfaces/ISecretAgent.ts
@@ -1,36 +1,20 @@
-import IConfigureOptions from "@secret-agent/core-interfaces/IConfigureOptions";
-import { IRenderingOption } from "@secret-agent/core-interfaces/ITabOptions";
-import IUserProfile from "@secret-agent/core-interfaces/IUserProfile";
-import SuperDocument from "awaited-dom/impl/super-klasses/SuperDocument";
-import { ILocationTrigger } from "@secret-agent/core-interfaces/Location";
-import ISessionOptions from "@secret-agent/core-interfaces/ISessionOptions";
-import { ISuperElement } from "awaited-dom/base/interfaces/super";
-import IWaitForResourceOptions from "@secret-agent/core-interfaces/IWaitForResourceOptions";
-import Response from "awaited-dom/impl/official-klasses/Response";
-import { IRequestInit } from "awaited-dom/base/interfaces/official";
-import Request from "awaited-dom/impl/official-klasses/Request";
-import IWaitForElementOptions from "@secret-agent/core-interfaces/IWaitForElementOptions";
-import Resource from "../lib/Resource";
-import IInteractions, { IMousePosition, ITypeInteraction } from "./IInteractions";
-import IWaitForResourceFilter from "./IWaitForResourceFilter";
-import ICreateSecretAgentOptions from "./ICreateSecretAgentOptions";
-import Tab from "../lib/Tab";
-import IAwaitedEventTarget from "./IAwaitedEventTarget";
+import IUserProfile from '@secret-agent/core-interfaces/IUserProfile';
+import SuperDocument from 'awaited-dom/impl/super-klasses/SuperDocument';
+import { ILocationTrigger } from '@secret-agent/core-interfaces/Location';
+import ISessionOptions from '@secret-agent/core-interfaces/ISessionOptions';
+import { ISuperElement } from 'awaited-dom/base/interfaces/super';
+import IWaitForResourceOptions from '@secret-agent/core-interfaces/IWaitForResourceOptions';
+import Response from 'awaited-dom/impl/official-klasses/Response';
+import { IRequestInit } from 'awaited-dom/base/interfaces/official';
+import Request from 'awaited-dom/impl/official-klasses/Request';
+import IWaitForElementOptions from '@secret-agent/core-interfaces/IWaitForElementOptions';
+import Resource from '../lib/Resource';
+import IInteractions, { IMousePosition, ITypeInteraction } from './IInteractions';
+import IWaitForResourceFilter from './IWaitForResourceFilter';
+import Tab from '../lib/Tab';
+import IAwaitedEventTarget from './IAwaitedEventTarget';
 
-export interface ISecretAgentConfigureOptions extends IConfigureOptions {
-  defaultRenderingOptions: IRenderingOption[];
-  defaultUserProfile: IUserProfile;
-}
-
-export default interface ISecretAgentClass {
-  // private options: ISecretAgentConfigureOptions;
-  configure(options: Partial<ISecretAgentConfigureOptions>): Promise<void>;
-  prewarm(options?: Partial<ISecretAgentConfigureOptions>): Promise<void>;
-  shutdown(): Promise<void>;
-  new (options?: ICreateSecretAgentOptions): ISecretAgent;
-}
-
-export interface ISecretAgent extends IAwaitedEventTarget<ISecretAgentEvents> {
+export default interface ISecretAgent extends IAwaitedEventTarget<ISecretAgentEvents> {
   readonly document: SuperDocument;
   sessionId: Promise<string>;
   Request: typeof Request;
@@ -80,6 +64,3 @@ interface IResolvedSecretAgent extends Omit<ISecretAgent, 'then'> {}
 export interface ISecretAgentEvents {
   close: ISecretAgent;
 }
-// hacky way to check the class implements statics we need
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export function SecretAgentStatics(constructor: ISecretAgentClass) {}
